requirements.txt
python-telegram-bot==20.4
psycopg2-binary
requests